requirements.txt
absl-py==2.1.0
attrs==24.2.0
Cartopy @ file:///home/user/zsy/Cartopy-0.24.1-cp310-cp310-manylinux_2_17_x86_64.manylinux2014_x86_64.whl#sha256=54f4d23961e0f9436baaf4747928361ccdcc893fa9b7bad9f615551bc8aa3fe8
certifi==2024.8.30
charset-normalizer==3.4.0
contourpy==1.3.1
cycler==0.12.1
exceptiongroup==1.2.2
fonttools==4.54.1
grpcio==1.68.1
h11==0.14.0
idna==3.10
joblib==1.4.2
kiwisolver==1.4.7
llvmlite==0.44.0
Markdown==3.7
MarkupSafe==3.0.2
matplotlib==3.9.2
numba==0.61.0
numpy @ file:///home/user/zsy/numpy-1.26.1-cp310-cp310-manylinux_2_17_x86_64.manylinux2014_x86_64.whl#sha256=8ab9163ca8aeb7fd32fe93866490654d2f7dda4e61bc6297bf72ce07fdc02f67
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
Pillow==9.5.0
protobuf==5.29.1
pynndescent==0.5.13
pyparsing==3.2.0
pyproj @ file:///home/user/zsy/pyproj-3.7.0-cp310-cp310-manylinux_2_17_x86_64.manylinux2014_x86_64.whl#sha256=ee60895f60cbd1a9c903ab2bc22adea63004296a1c28b8775a11cf50905cf085
pyshp==2.3.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.32.3
scikit-learn==1.5.2
scipy==1.14.1
selenium==4.26.1
shapely @ file:///home/user/zsy/shapely-2.0.6-cp310-cp310-manylinux_2_17_x86_64.manylinux2014_x86_64.whl#sha256=b3304883bd82d44be1b27a9d17f1167fda8c7f5a02a897958d86c59ec69b705e
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
threadpoolctl==3.5.0
torch==1.12.1+cu116
torchaudio==0.12.1+cu116
torchvision==0.13.1+cu116
tqdm==4.67.0
trio==0.27.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.2
umap==0.1.1
umap-learn==0.5.7
urllib3==2.2.3
websocket-client==1.8.0
Werkzeug==3.1.3
wsproto==1.2.0
